Minimal version ability to restart simulation at new p (on same grid)...
We still require the user to specify RESTART_FROM_AUX and AUX_ORDER in
the run configuration file.  The AUX_ORDER restart file is read and
the converved variables POL_ORDER solution is constructed by
interpolating the AUX_ORDER conserved variables using built-in mfem
capability, specifically a call to GridFunction::ProjectGridFunction.
The next time a restart file is written, it will overwrite the
AUX_ORDER restart with the new order restart.  The user is responsible
for maintaining a copy of the restart if that is desired.