Initial running version of quasi-magnetostatic solver (#68)
The solver itself is housed in the QuasiMagnetostaticSolver class,
which provides public member functions to initialize mesh, required
spaces, and curl-curl operator (Initialize()), initialize the right
hand side (InitializeCurrent()), and solve the system (Solve()).  This
functionality is implemented in src/quasimagnetostatic.{hpp, cpp} and
is called from main when the --em-only option is provided.

Note that this commit includes the files pfem_extras.{hpp, cpp} which
are from MFEM.  These encapsulate some capabilities (e.g., the
DivergenceFreeProjector class) used in the EM code and are not
directly provide through the library (as of mfem v4.2 at least).
These files are housed in a separate subdir from the rest of the
source, which required using the 'subdir-objects' option in
AM_INIT_AUTOMAKE in configure.ac.