Handle discontinuous source current fcn consistently in serial and parallel
Previous call to ParGridFunction::ProjectCoefficient (in
QuasiMagnetostaticSolver::InitializeCurrent) can lead to different
results in serial and parallel, because DoFs that are shared between
elements may not be initialized the same way.  In MFEM, this is
handled by calling ParGridFunction::ProjectDiscCoefficient, which
checks the element attribute for each DoF and uses the element with
the largest attribute to initialize the DoF.  This is great, but the
function seems to not work for Nedelec elements because of the dof
numbering scheme (specifically FiniteElementSpace::GetElementVDofs can
return negative DoF indices for Nedelec elements, and these raw
indices are then used in index into mfem::Array objects, which is
illegal).  Here, we fix this in tps code with the functions
{Local,Global}ProjectDiscCoefficient in src/utils.{hpp,cpp}.