WIP (#75): Starting full Maxwell implementation
Start by making a class called SeqsMaxwellFrequencySolver that will
house the new capability.  For now, the class initializes and solves
just the governing equations for the electric potential (i.e., the
SEQS problem in the parlance of Ostrowski and Hiptmair.  There is no
testing yet, so very preliminary... probably buggy.