WIP (#75): Add E, B, and I calculations
First, compute E and B, since these are the physically meaningful
quantities.  Specifically, as a postprocessing step, we compute

B = curl(A)
E = -grad(phi) - j*omega*A

where phi and A denote the "full" or "total" potentials, which are
split up by Ostrowski and Hiptmair.

Only E and B are written to the paraview output now.

Regarding the current, it is split into two parts, the "static" and
"induced" parts, as defined by Ostrowski and Hiptmair, and evaluated
accordingly.