Update reference solutions for axisymmetric annulus and pipe
Necessary due to change in isothermal wall BC.  In original
implementation, we used the viscous flux evaluated at the wall state.
In the implementation introduced in #100, we use the average of the
viscous flux at the wall state and at the interior state.  The latter
is more consistent with the handling of the viscous flux elsewhere, so
we choose it.