Baby steps to a coupled problem (#81)
Here we develop a trivial example of a 'coupled' solver.
Specifically, we introduce a new class (IndependentCoupling) that
derives from TPS::Solver and contains both a flow solver and an EM
solver.  So, although no information is exchanged (i.e., there is no
mathematical coupling) we can at least solve a flow and an EM problem
in a single run of tps.