Fixes for build with gslib support
Because of MFEM_CXXFLAGS and MFEM_LIBS variables, when mfem is built
with gslib, the necessary include and lib info automatically
propagate, so we no longer need to explicitly provide gslib info.

Also, because of changes in the M2ulPhyS class (e.g., the pointer to a
TPS::Tps class must be passed through the ctor), the interpolation
utility does not build.  It has been removed temporarily.