Implement inviscid flux Jacobian for element interior integrals
To make this easy, we have implemented this the 'traditional' way
(i.e., by building a NonlinearFormIntegrator that integrates <grad v,
F> and its Jacobian for each element), which allows us to take
advantage of MFEM's assembly loops.  However, since the original tps
implementation builds a linear operator correspondingt <grad v, v> and
then applied it to the flux vector, we have had to reimplement the
residual calculation, not just add the Jacobian.  For now, this is
only done for Euler.  By default, the residual is built with the
original scheme.  To invoke the new method, use flow/useNodalInterior
= 0 in your input file.