Reset dev in device_.Configure to be consistent with original
Specifically, we need the second argument passed to
mfem::Device::Configure to be the index of the gpu for that mpi rank
to use.  When multiple gpus are available this should run from 0 to
(number of gpus per node - 1).  This was the original behavior, but
somewhere it was changed to where every mpi rank would get the same
dev argument, meaning that only 1 gpu on a node could be used.