Add configure option to build _GPU_ code path on cpu
Even when gpu(s) aren't available, the MFEM macros we are using to
build the _GPU_ code path provide valid code.  So there is no
fundamental reason that the gpu code path can't build and run on a cpu
also.  The current code does build but gives incorrect results because
some pieces of the gpu code path assume parallelism that doesn't exist
on the cpu.