For cuda build, move -dc into AM_CPPFLAGS in src/Makefile.am
This flag was previously part of CUDA_CXXFLAGS in configure.ac, but
that approach leads to failure in AX_CXX_CHECK_LIB in mfem.m4.  This
failure occurs b/c -dc tells nvcc just to compile rather than compile
and link, so that the output of the nvcc test program build is not
executable, which leads to failure in AC_LINK_IFELSE.

By moving -dc into AM_CPPFLAGS in src/Makefile.am (just for cuda
builds of course) we avoid this problem.