Hacky fix for preprocessor macro redefined warnings
tps_config.h defines a number of macros (e.g., PACKAGE_NAME) that are
also defined by HYPRE_config.h, which is pulled in via mfem.  This
leads to a ton of warning messages that obscure other warnings from
the compiler.  As far as I can tell, there is no way to disable just
this warning.  For instance, see [this
post](https://stackoverflow.com/questions/8159978/temporarily-disable-gcc-warning-on-redefinition)
which is ancient, but apparently still true.  The fix here is to wrap
our include of mfem.hpp in a separate header called tps_mfem_wrap.hpp,
which #undefs the macros we directly control.

Note that this does not work for macros that tps does not directly
control.  Specifically, HYPRE_TIMING can still lead to macro redefined
warnings (depending on your hypre configuration), but that is entirely
internal to mfem and I don't see any way to change the situation
without modifying mfem or hypre.