Refactor s.t. hip version works
Two changes:

1) Don't use new in device code.  It isn't supported by hip.  We could
replace with malloc (since hip supports that, at least with
-D__HIP_ENABLE_DEVICE_MALLOC__), but malloc/new in device functions
probably isn't great for performance anyway.  For now, we replace with
regular compile-time-sized C arrays.  The sizes will need to be
changed in the future to support additional eqns (but that has to
happen in lots of places in the _GPU_ code path anyway).

2) Replace the arguments passed to RiemannSolver::Eval_LF in
DGNonLinearForm::evalFaceFlux_gpu.  Specifically, instead of u1, u2,
and nor (which are local arrays), pass pointers into d_uk_el1,
d_uk_el2, and d_shapeWnor1, which are pointers to mfem::Vector on the
device.

It is unclear to me why change 2 should matter, and if we are only
using non-virtual functions, it does not.  But, for some reason, the
code produces garbage without this change.  Maybe we are somehow
running afoul of the limitations of hip's virtual function support
(see section 1.3.13.4.1.6 on page 60 of the ROCm Documentation,
Release 4.5.0: https://rocmdocs.amd.com/_/downloads/en/latest/pdf).