Update EM preconditioning (to allow user to add background conductivity)
This helps substantially for problems where the conductivity is large
in one portion of the domain and zero on another.  When the
conductivity was zero everywhere, the HypreAMS preconditioner (using
SetSingularProblem) is working very well.  However, when adding
conductivity in one part of the domain (and still using
SetSingularProblem), the convergence tends to stall.  Adding a small
background conductivity everywhere (just to the preconditioner
operator, not the PDE itself) and not using SetSingularProblem (since
the preconditioner is positive definite in this case) eliminates this
problem.