Add gslib support to MFEM build in quartz container
This change adds gslib support, which is necessary to use tps
interpolation routines, to the mfem build in the quartz container.  In
addition, we switch to the mfem version being used in the tps_env
container.  This change provides support for gslib interpolation on
mixed meshes.

In making this change, the petsc dependency in mfem was dropped.  When
attempting to use the mvapich2-based petsc module from ohpc, the mfem
build complains with the following message:

PETSc was configured with one MVAPICH2 mpi.h version but now appears
to be compiling using a different MVAPICH2 mpi.h version

This appears to be saying that the mvapich2 used to build the petsc
module is different from that installed by `yum install mvapich2-psm2-gnu9-ohpc`.
I don't see an easy workaround (i.e., one that doesn't require a new
ohpc package) for this, but tps currently doesn't use petsc through
mfem (or otherwise) anyway.