Make it possible to pass input filename to tps::ParseInput
This enables variant where we parse the command line inputs directly
and then set up tps, which is useful in some instances.  Specifically,
it is useful to bring the field interpolator utility (which
interpolates a restart state to a new mesh) back to life, since it
requires that two input files be available on the command line, where
tps::ParseCommandLineArgs only handles one.  Now we can handle the
command line parsing externally to tps but still parse tps input
files by passing them.