Minimal fixes to get gpu code path to build with nscbc branch mods
This does not imply that the functionality implemented for the CPU on
the nscbc branch works on the GPU.  That is explicitly not the case.
These mods revert changes to the GPU code path that are causing it not
to compile.  By doing so, the capabilities the changes support (e.g.,
position dependent BCs) are definitely not supported on the GPU.