Refactor tests on zaphod
The gpu in zaphod (AMD Radeon Pro WX 3200, gfx803) is not fully
supported by rocm 5.X.  See
https://docs.amd.com/bundle/Hardware_and_Software_Reference_Guide/page/Hardware_and_Software_Support.html
(updated on March 7, 2022 and still the latest as of this commit).
While the initial, static fcn only implementation as well as the more
recent hybrid approach (with some _HIP_ specific code, e.g. for the
viscous fluxes) both passed the tps regression test suite on this
card, the refactored code that unifies the hip and cuda paths leads to
failure.  However, the tests pass on other amd gpus (gfx90a).

Given this behavior, this commit changes the zaphod tests to run the
"gpu-cpu" variant---i.e., the _GPU_ code path, but running on the cpu.