Fixes for LTE transport
First, evaluate electrical conductivity in LTE transport.  Second,
make sure electron thermal conductivity is initialized (to zero) by
LTE transport.  The reason to set it to zero is that it is already
incorporated into the "total" conductivity that is used to set the
heavy species conductivity.  Since the flux function adds the electron
conductivity (in the single T model), we must set it to zero to avoid
either using potentially uninitialized space or double counting a
piece of the conductivity.