Fix variable order restart calculation of primitive variables
For a variable order restart, in M2ulPhyS::restart_files_hdf5, the
primitive variable array was initialized incorrectly.  In a practical
simulation, this is irrelevant, since the primitives are computed
prior to the evaluation of the right-hand-side.  But, if we are going
to bother to do it (e.g., in case a visualization file is written
before a time step is taken), the calculation should be correct.