Modifications to avoid issues reported in #184
This commit resolves the problems reported in #184.  In particular, we
make the computeEnergySink method virtual, which is necessary for the
code to work as expected, and eliminate the problems observed in the
CUDA branch by changing `necTable_` from `TableInterpolator *` to
`LinearTable`.  This also eliminates the `new` in the `NetEmission`
ctor, which will be necessary for porting to HIP.

However, the root of the problem described in #184 is still unclear to
me, and we will have to revist if we ever want to generalize beyond
`LinearTable` here.