Require (energy, density) -> T table for LTE table look-up
This table is used to generate an initial guess when inverting the
mapping from (temperature, density) to energy.  In principle it is not
necessary (or could be generated from the (T, rho) -> e data).  But,
in practice, it is easily available, so we might as well take
advantage of it.