Some clean up in volumeFaceIntegrationArrays
Overall idea is to replace arrays that currently hold more than one
thing with multiple arrays, for clarity, and to rename some objects,
also for clarity.

For instance, here we start to replace posDofIds with
element_dof_offset and element_dof_number.  But, for now, I only
eliminate posDofIds in DGNonlinearForm, because posDofIds gets passed
to BC evaluation methods, which will require a more extensive refactor
that I'll return to later.