Replace volumeFaceIntegrationArrays::shapeWnor1 with multiple arrays
For clarity, I prefer not to store different quantities as part of
same array unless there is a pressing need to do so, and there doesn't
appear to be such a need here.  So, this commit split out shape fcns,
quad weights, and normals that are currently in shapeWnor1 into
separate vectors, specifically, face_el1_shape, face_quad_weight, and
face_normal.