Further clean-up of volumeFaceIntegrationArrays
In this commit, we split and rename volumeFaceIntegrationArrays.  It
is split into two pieces that logically hold different information
(i.e., information about elements and information about faces).  Each
piece is given its own struct and these structs are data members of a
new struct (precomputedIntegrationData), which replaces
volumeFaceIntegrationArrays.

To goals in doing this: 1) clearer and more readable code, and 2)
enable further clean up of the BC interfaces that use this data
s.t. posDofIds can be eliminated.