Prototype new struct to store and pass quadrature point data
Changing (e.g., adding the position as an input) inputs to flux and
transport calculations is currently very cumbersome because every call
to these functions has to be updated when the inputs change.  To avoid
this, we can pass a struct with all the necessary info and then just
update the struct.