Workaround cuda weirdness with virtual fcns that call virtual fcns (#108)
TransportProperties::GetViscosities exposed a weirdness in cuda for
virtual functions that call virtual functions.  It *seems* (I will try
to produce a small example elsewhere) that for a virtual function that
calls a virtual function with the same signature (or not even
necessarily exactly the same but close, e.g., different only by
const), nvcc cannot determine that the function is not called
recursively.  For recursive functions, nvcc can't appropriately size
the stack, leading to warnings like the following

nvlink warning : Stack size for entry function '_ZN3gpu19freeDeviceRadiationEP9Radiation' cannot be statically determined

Sometimes (as with the destructor above) this appears to be harmless,
but sometimes, probably when the default stack isn't large enough, it
leads to test failures.  GetViscosities exposes this problem because
in the argon transport classes, it calls
GasMixture::computeSpeciesPrimitives, which appears to have similar
enough signature such that nvcc thought it might be recursion (as can
by calling nvlink with the --dump-callgraph option to see what it
thinks the possible calls from GetViscosities are, and indeed, for
ArgonMinimalTransport for instance, it thinks GetViscosities might be
calling itself, even though it definitely doesn't).

Regardless, as a work around for now, I have modified the signature of
GetViscosities to avoid the issue.