Add variant of sgsSigma that uses lapack for singular values
While porting the sigma model to the gpu, I realized that the
sensitivity of the sigma model test case to order of operations
changes (e.g., noted in commit f19a522) is likely due to the way the
singular values are being computed.  Specifically, the model requires
the singular values of the velocity gradient, g, which are evaluated
by taking the square root of the eigenvalues of g^T * g.  This
approach can be much more sensitive to perturbations in g than other
algorithms for computing the singular values.

Since mfem provides an interface to lapack to get singular values as
part of the DenseMatrix class, this commit adds support for this
approach.  However, it is only available currently on the cpu and when
mfem was built with lapack support.  For an exactly analogous approach
to work for gpu builds, we need a __device__ interface to a
lapack-like library, but I have not found such a capability.  For
example, cuSolver provides the SVD, but not as a __device__ function,
as far as I can tell.  We would need to refactor substantially to make
use of this.  So, for now, on the device, we use the original approach.