Loosen tolerance on sgsSigma.test
Due to sensitivity of singular value calculation, have to relax the
tolerance for different code paths to pass the test with same
reference solution.  I have confirmed that this sensitivity goes away
if we replace the current singular value calucation with lapack, but
this is not yet implemented on all paths.  So, for the moment at
least, I'm relaxing the tolerance.