Update mesh spacing passed to subgrid models
Two changes:
* for face terms, pass the local mesh spacing from the each side of the face
* use h/p rather than just h

Note that for shared faces, we have to be a bit careful, since
Mesh::GetElementSize does not work for elements that are not owned by
the local mpi rank.  Instead we compute the element size from the
Jacobian of the ref to physical space mapping, which is available.

Shared and boundary face element spacing fixes for sgs models

1) divide by p on boundary on cpu path
2) handle shared and boundary faces on gpu path