Update input file used by interp_em.test
This test was failing b/c an input to the updated viscosity sponge was
not present in the input file.  But, that capability is irrelevant to
the test, so this commit just changes the input file to turn it off.

NB: This commit builds and passes all regression tests on the cpu
path.  The gpu path however does not build yet.