Use slip bc in plate.test
This requires regenerating the reference solution, since the original
reference solution was generated with a version of the code where, b/c
of a bug, the slip bc was not used and the inviscid bc was applied
instead.  In addition, we make the minimal edit to the gpu path to
make the test pass there as well.