Update mfem for hip build on zaphod
Also updated zaphod to rocm 5.6.  With this new version, the hip build
died, complaining that we were using __host__ sqrt in a __device__
function and pointing to the initialization of the Chapman-Enskog
coefficients (e.g., viscosityFactor_) in argon_transport.hpp.  Moving
those initializations into the body of constructor for the
ArgonMinimalTransport class avoids the problem.