Do not build mixing length transport for gpu
Specifically for nvcc, it seems that building the
MixingLengthTransport class, even when it isn't used, confuses the
compiler, leading to many 'Stack size for entry function... cannot be
statically determined' warnings and subsequent test failures.  I
*guess* this may be due to the fact that this class contains a pointer
to a TransportProperties object, which makes it hard for nvcc to
figure out that nothing is actually recursive?

Anyway, since we're not ready to support the gpu right now, just #if
it out.