Update coupled-3d reference solution
B/c there is now a check for negative Joule heating where the forcing
is added, this test has changed slightly.  B/c of interpolation, the
Joule heating can go slightly negative on the flow mesh, which is
happening in this case.  Previously these negative values were added.
Now, the Joule heating is only added if it is non-negative.  This
modification changes the solution, leading to this reference solution
update.