Reduce number of fe spaces to three
For simplicity during development and debugging.  The minimum number
of spaces we require is three: a dim-dimensional space for the
velocity, a space for scalar variables (e.g., pressure), and a
num_equation-dimensional space for the whole solution.  In the future
may wish to expand to have different spaces for different scalar
variables, but not yet.