Switch to preserving pressure exactly in LTE -> non-equilibrium restart
Because of differences between the excited states (both neutral and
ion) between the offline equilibrium calculations that underly
LTE-based runs and those included explicitly in TPS non-equilibrium
reacting simulations, it is in general not possible to find an
equilibrium state (i.e., a TPS state vector) that exactly matches the
pressure, density, and temperature simultaneously from the LTE result
in the non-equilibrium model.  Here, we choose to match pressure and
temperature, which leads to a slight perturbation in the overall
mixture mass density.