A function to evaluate the electrical conductivity on the grid
This function allows us to update the electrical conductivity just
prior to EM solve, which is required for proper restarts.  Also, we
include a flag enabling the user to skip this update.  This can be
useful for ramping up cases by generating a Joule heating based on an
artificial electrical conductivity field.