Work towards enabling use of MixingLengthTransport on device
The goal is to enable MixingLengthTransport on the gpu.  Prior to this
commit, this class was surrounded by a #if that prevented it from
being compiled in gpu builds.  Simply removing this #if leads to test
failures, b/c the design of the MixingLengthTransport---specifically,
where some methods call the same method from a different class.  This
seems to confuse nvcc, even though the MixingLengthTransport class is
not even instantiated in the gpu path yet.  Specifically, test
failures are observed (e.g., the parallel tests in cyl3d.gpu.test
fail).

I think the problem is related to the issue first noticed in #184 and
then dealt with (via a workaround) in #203.  To work around it here, a
new class MolecularTransport is introduced, where the flux and source
transport functions can have a different signature (b/c they never
need the distance function, for example).  This is enough to avoid the
problem---i.e., all the tests pass.