Instantiate MixingLengthTransport on the device
Wanted to use dynamic_cast to cast the TransportProperties pointer
that M2ulPhyS holds to MolecularTransport (as in the cpu path), but
host-side dynamic_cast on a device-side pointer seg faults, and
device-side call to dynamic_cast isn't allowed.  So... go ahead and
pass the TransportProperties pointer and use a regular cast on the
device.