Distance function fixes for the gpu path
Two bugs in setting the distance function arrays for the interior
faces on the gpu path.  This data are only used in the
MixingLengthTransport class, and since we only just started trying to
make that work on the gpu path, it was just noticed now.

The useBCinGrad flag is not implemented on the gpu yet, so
pipe.mix.test fails.  However, if that flag is set to false on the cpu
path, then the result of that case is the same on the gpu and the cpu
(to within the soln_differ tolerance).