Towards useBCinGrad flag on device path
This commit implements the useBCinGrad flag on the gpu path.  When
this flag is on, for a no-slip adiabatic wall, the velocity (0) and
user-set temperature are used in boundary terms that appear in the
system defining the gradient.  Otherwise, these terms are ignored,
which is incorrect.

Since the terms were ignored before, there was no infrastructure for
looping over boundary faces in forming the gradient system on the gpu
path.  This commit adds this infrastructure, which is slightly more
complicated than analogous face loops (i.e., interior and 'shared'
faces) due to the fact that periodic faces are counted as boundary
faces by mfem (see mfem::Mesh::GetNBE vs mfem::Mesh::GetNFbyType).

The infrastructure is added in the minimal way possible---i.e., just
for no-slip, isothermal walls.  The BC infrastructure will need to be
refactored to make it easy to support all BC types.