Fix issue with negative number densities.
Negative values can occur due to rounding and/or interpolation error,
among other things.  So, when negatives are detected, we force this
value to zero in face_integrator.cpp and BCintegrator.cpp. We also
check for negative values at the end of the iteration with
Check_undershoot.  Undershoot can occur due to discretisation error in
the solution of the transport equation. The use of a TVD scheme could
mitigate this effect.

This commit also adds variables for the temperature indices.