Split M2ulPhyS::restart_files_hdf5 into multiple functions (#244)
This is an intermediate commit.  The idea is to extract the most of
the read and write functionality from M2ulPhyS::restart_files_hdf5
into separate functions.  In the previous state the read and write
code were intertwined, which made it difficult to understand.  The
eventual goal is to move the parts that are generic into
IODataOrganizer (and related classes), in order to minimize the
restart read/write code within M2ulPhyS, which should make it easier
to reuse these capabilities for other Solver types down the road.