Move primary data read support into IODataOrganizer class (#244)
This is mostly analogous to the write support from the previous
commit, with new read functions being added to IODataOrganizer.

But, in addition we have to support the option to read data
from a different order solution (to support restarting from lower
order fields).  This is facilitated by adding "aux" objects
(FiniteElementCollection FiniteElementSpace, and ParGridFunction) to
the IOFamily class.  In the previous implementation, these were
instantiated directly within the read function and were specific to
the DG discretization used by M2ulPhyS.  The new implementation should
be generic (although that generality is not yet tested).