added support for restarting RMS stats, fixed small bug in vector-vector product, switched to mass-based thermoP
trevilo note: this commit was cherry-picked from e9b8bbf on
dev-loMach-variableCoeffPP.  Edits in the original commit to io.hpp
were discarded.  Edits in the original commit to io.cpp were modified
b/c of the the io structure.  Edits in the original commit to
averaging_and_rms.cpp were modified to deal with conflicts.