Initial version of interface for low Mach development
In developing the low Mach solver, it will be useful to have base
classes for different component models.  This commit includes draft
base classes for the thermochemistry, which is responsible for the
energy and species equations, and for the flow, which is responsible
for the velocity field.  The only thing that is important about these
base classes is that they codify the information that each model must
supply to the other.  To make the interface somewhat extensible,
fields are "exchanged" through structs that hold pointers to all the
necessary fields.

The interface is tested via a trival program
(test/test_split_interface.cpp) which just instantiates the simplest
versions of the model classes and ensures that the data obtained via
the interface is as expected.