Disable averaging in LoMachSolver (temporarily)
The Averaging class introduces a bunch of cruft (e.g., fe spaces and
functions) into LoMachSolver that shouldn't really be there.  This is
because Averaging is written in a brittle way that really only
supports the compressible flow code.  So, it ought to be refactored.
In the meantime, disable it in LoMachSolver to keep things clean.