Remove LoMachSolver::parseBCInputs
This fcn was parsing bc info into config for use in low Mach solvers.
But, goal is to eventually eliminate config from LoMachSolver.  Plus
this function was build to support BC types that don't necessarily
make sense in the low Mach context.  Instead, in the LoMachSolver, the
classes that get instantiated and own the state (e.g., ThermoChem,
Tomboulides) will parse BC data and figure out how to use it.