Eliminate RunConfiguration input to partitioning_file_hdf5
This input was passed so that we could get the base filename for the
partitioning file.  However, this defaulted to "partition" and no
input file mechanism was ever provided to change it from that default.
So, now the default is simply hardcoded in partitioning_file_hdf5,
which eliminates the need for RunConfiguration to be passed.