Baby steps toward generalizing Averaging (#247)
This commit introduces AveragingFamily, which has a pointer to an
instantaneous field (not owned) which contains variables being
averaged and pointers to mean and rms fields (owned).  When
registerField is called a new AveragingFamily is added to a
std::vector held in Averaging.  For now, the internals are still only
correct for the M2ulPhyS case (e.g., the rms calc is specific to how
the data are stored there).