Add regression test based on Le Quere benchmark
There are known problems with this test due to restart issues in the
closed domain case (specifically, since P and dP/dt aren't stored in
the restart, running 2 steps is not the same as running 1, restarting,
and running another, even with first order bdf).

Nonetheless, having some regression coverage here is beneficial as we
develop other capabilities, to ensure the variable density Tomboulides
doesn't change unexpectedly.