Initialize electrical conductivity
NB: This initialization does not matter if the input deck has

cycle-avg-joule-coupled/fixed-conductivity = False

because the conductivity will be overwritten by the conductivity
computed from the solution.  However, for "start-up" (e.g., when
heating the flow artificially to initialize a plasma simulation), it
can be useful to have a user-specified conductivity field.  This
commit provides a (hardcoded for now) function for this purpose.