Exchange face neighbor data before we compute primitives
After restart, make sure all data valid before using it.  I don't
think this could ever cause an incorrect result for a time step b/c
the info should be exchanged prior to all calculations that influence
the step.  But, it can lead to (and has recently led to) to failing
asserts in the primitive variable calculations.