Ensure SpongeZone::sigma is properly initialized
In the SpongeZone ctor we have

sigma = new ParGridFunction(&fes);
*sigma = 0.;
double *hSigma = sigma->HostWrite();

and the subsequent loop (prior to this commit) only set some entries
in hSigma.  On systems where a device is available, this process can
lead to uninitialized values in the sigma field, b/c *sigma = 0
initializes the device memory and `sigma->HostWrite` immediately
invalidates that (i.e., no copy is done) but then hSigma isn't fully
initialized.

In this commit, we eliminate the *sigma = 0. initialization in favor
of setting all entries within the loop.