Enable singleFileWrite mode on low Mach path
The serialized restart capabilities have not been implemented for the
low Mach path.  This commit begins to develop this capability by
enabling singleFileWrite mode.  To do so, we just need to fill the
array mapping the local element index to the global element, which is
used when serializing the data.  This commit adds the necessary bit of
code to fill that array into the MeshBase class.