Fix bug in ReactingFlow::speciesProduction
The primitive state was being filled with mass fractions for the
species entries and then passed to GetConservativesFromPrimitives, but
this function expects that the species entries of the primitive state
contain molar densities.  Because of this, the returned conserved
state was incorrect and the resulting molar densities returned by
computeNumberDensities were also incorrect, which then made the
species production terms incorrect.