Bug fixes in ReactingFlow::heatOfFormation
Three bug fixes here:

1) The units of the formation energy from the input file is J / mol,
so we need to divide by the species weight (kg / mol) here to get the
formation energy per unit mass.

2) The vector hw_ was sizes to have entries for all species and the
contributions were not summed.  We need hw_ to be a scalar field with
the contributions from all species summed.

3) Sign error

Note that this commit does not add the sensible enthalpy contribution
(see Poinsot and Veynante, Theoretical and Numerical Combustion, 2005,
eq. 1.62).