Bug fixes in ReactingFlow
Three bugs were found and fixed while working on the low Mach analog
to argon_minimal.binary.test.  The fixes are

1) In ReactingFlow::speciesStep, set the initial guess for
the iterative linear solve to the mass fraction for the current
species from the previous time step.  Because of how Yn_next_gf_ is
managed, that was not the case and sometimes the electron species
solve in particular struggled because of it.

2) Multiply the diffusivities returned by
computeMixtureAverageDiffusivity by the mixture density (i.e., so that
the diffision term in the species equation is div( \rho D \grad Y) ).

3) Populate the species entries of the primitive state in
ReactingFlow::updateDiffusivity with the mole densities, rather than
the mass fraction.