Attempt to workaround gpu issues in axisymmetric Tomboulides
The u_next_rad_comp_gf_ ParGridFunction seems to be causing problems
on some gpu systems.  This is intended to alias the first
(i.e. radial) component of the velocity grid function.  It is
constructed via

u_next_rad_comp_gf_ = new ParGridFunction(pfes_, *u_next_gf_);

But, this leads to failures like the following:

Verification failed: (IsHostMemory(h_mt)) is false

in mfem::MFEM_VERIFY_TYPES, which is called (eventually) from
Vector::HostRead.  Initially I thought this had to do with a failure
to properly call Vector::SyncMemory, since the base memory (in
u_next_gf_) is being modified.  But, an attempt to use this had no
impact on the problem.

So, in this workaround, I eliminate the alias and simply copy the
first component into a new grid function.  This seems to work, but it
would be nice to understand the problem with the original approach,
b/c clearly we'd rather not copy when it shouldn't be necessary.