Initialize velocity gradient and subgrid viscosity
Specifically...

1) Make sure velocity gradients are properly initialized by
Tomboulides class, first to 0 in initializeSelf, and then by a call to
a evaluateVelocityGradients (a new function) in initializeOperators,
in order for the initial velocity gradients to be consistent with the
initial velocity field.

2) Call step in AlgebraicSubgridModels::initializeOperators, so that
the subgrid viscosity is computed based on whatever the velocity
gradients are at this point.

3) Change the order of initializeOperators calls in LoMachSolver, so
that the turbulence model is initialized after the flow, which ensures
that the velocity gradient will be valid when the subgrid viscosity is
initialized.