Update reference solutions
For tests affected by the fix to the turbulence model initialization.
Note that this affects lomach-lequere.test, even though it does not
use a turbulence model, b/c u_next_gf_ was not properly initialized
before this fix.  Specifically, while u_curr_gf_ was set from the
restart file, u_next_gf_ was not.  Thus, for the first the first
thermal system step, the incorrect velocity was used.