Add TpsVectorMassIntegrator to support partial assembly
This class derives from mfem::VectorMassIntegrator and overrides the
AssemblePA method to enable support of partial assembly of 'vector'
mass matrices with a scalar coefficient.  This capability is needed in
Tomboulides to build the density-weighted mass matrix on the velocity
space.

The current version of mfem::VectorMassIntegrator only supports
constant coefficients.  This extension to general scalar coefficients
should go back into mfem, but for initial prototyping and testing, we
put it into tps.