Switch from cuda-memcheck to compute-sanitizer
After OS upgrade on CI system, cuda-memcheck fails to detect the known
errors in badcuda.cpp, which are based on examples from the nvidia
docs:

https://docs.nvidia.com/cuda/archive/11.4.1/cuda-memcheck/index.html#example-use-of-memcheck

It isn't clear why this is the case, but since cuda-memcheck is
deprecated in favor of compute-sanitizer in later versions anyway, I
am switching to compute-sanitizer.  It successfully detects errors in
badcuda.