Make lomach restart test pass with flux
For some reason that I don't understand, lomach-flow.test 5 (i.e.,
"verify consistent serialized restart for Tomboulides with lid-driven
cavity with 2 mpi ranks") consistently fails on one of our testing
systems that uses flux b/c it thinks that the restart file written by
the solver does not exist when the test script attempts to mv the
file (the mv on line 133 of this version).

This commit, which adds some steps between the tps solve and the mv,
appears to avoid the issue.